anbapi/requirements.txt
fastapi
uvicorn[standard]
gunicorn
SQLAlchemy>=2.0
psycopg2-binary
pydantic
pydantic[email]
pydantic-settings
python-dotenv
rich
asgi-correlation-id
python-json-logger
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
celery
redis
slowapi             
prometheus-fastapi-instrumentator
faker
psycopg[binary]